Move replaceForWithForallIfNeeded() & helpers
