+++ compiler/resolution/Makefile.share
@@ -26,7 +26,6 @@ RESOLUTION_SRCS =                                              \
                   functionResolution.cpp                       \
                   generics.cpp                                 \
                   implementForallIntents.cpp                   \
-                  implementForallIntents2.cpp                  \
                   initializerResolution.cpp                    \
                   lateConstCheck.cpp                           \
                   lifetime.cpp                                 \
